Split up schema definition
